financial_portfolio_scanner/requirements.txt
# Financial Portfolio Scanner - Dependencies

# Microsoft Agent Framework for building AI agents
agent-framework>=1.0.0b251016

# OpenAI API integration for AI capabilities
openai>=1.54.0

# Environment variable management for configuration
python-dotenv>=1.0.0

# Data manipulation and CSV export functionality
pandas>=2.1.0

# Rate limiting for API calls
asyncio-throttle>=1.0.2

# HTTP requests for API integration
requests>=2.31.0

# For optional visualizations and charts
matplotlib>=3.7.0

# For console output formatting
tabulate>=0.9.0

# Testing framework
pytest>=6.2.5
pytest-cov>=2.12.0